Mostly rearranging. Renamed Client to GremlinConnection
